Prep for protein user schemas migration
