+++ ms2/src/org/labkey/ms2/protein/query/CustomAnnotationSchema.java
@@ -20,26 +20,23 @@ import org.labkey.api.data.Container;
 import org.labkey.api.data.ContainerFilter;
 import org.labkey.api.data.TableInfo;
 import org.labkey.api.module.Module;
+import org.labkey.api.protein.CustomAnnotationSet;
+import org.labkey.api.protein.CustomAnnotationSetManager;
 import org.labkey.api.protein.ProteinSchema;
 import org.labkey.api.query.DefaultSchema;
 import org.labkey.api.query.QuerySchema;
 import org.labkey.api.query.UserSchema;
 import org.labkey.api.security.User;
-import org.labkey.api.protein.CustomAnnotationSet;
-import org.labkey.ms2.protein.ProteinManager;
 
 import java.util.Map;
 import java.util.Set;
 
-/**
- * User: jeckels
- * Date: Apr 3, 2007
- */
 public class CustomAnnotationSchema extends UserSchema
 {
     public static final String SCHEMA_WITHOUT_SEQUENCES_NAME = "CustomProteinAnnotations";
     public static final String SCHEMA_WITH_SEQUENCES_NAME = "CustomProteinAnnotationsWithSequences";
     public static final String SCHEMA_DESCR = "Contains data about custom protein annotations.";
+
     private final boolean _includeSequences;
 
     public static void register(Module module)
@@ -61,13 +58,13 @@ public class CustomAnnotationSchema extends UserSchema
         });
     }
 
-    private Map<String, CustomAnnotationSet> _annotationSets;
+    private final Map<String, CustomAnnotationSet> _annotationSets;
 
     public CustomAnnotationSchema(User user, Container container, boolean includeSequences)
     {
         super(includeSequences ? SCHEMA_WITH_SEQUENCES_NAME : SCHEMA_WITHOUT_SEQUENCES_NAME, SCHEMA_DESCR, user, container, ProteinSchema.getSchema());
         _includeSequences = includeSequences;
-        _annotationSets = ProteinManager.getCustomAnnotationSets(container, true);
+        _annotationSets = CustomAnnotationSetManager.getCustomAnnotationSets(container, true);
     }
 
     protected Map<String, CustomAnnotationSet> getAnnotationSets()
